requirements.txt.txt
openai==0.28.1
python-dotenv
streamlit>=1.30.0